Only install chruby if it isn't installed.

+++ bin/install.sh
@@ -6,6 +6,8 @@ command -v apt-get >/dev/null 2>&1 || { echo >&2 "apt-get is required"; exit 1; 
 command -v wget >/dev/null 2>&1 || { echo >&2 "wget is required"; exit 1; }
 command -v make >/dev/null 2>&1 || { sudo apt-get install make; };
   # make is a prereq for ruby-install and chruby
+sudo apt-get update
+sudo apt-get install --assume-yes mariadb-server mariadb-client ruby-mysql libmysqlclient-dev
 
 if !(command -v ruby-install >/dev/null 2>&1)
 then
@@ -29,24 +31,29 @@ fi
 
 if !( command -v chruby >/dev/null 2>&1)
 then
-  # Change to a temporary directory so stuff isn't cluttered
-  pushd .
-  cd /tmp
+  
+  if !(test -e /usr/local/share/chruby/chruby.sh)
+  then
+    # Change to a temporary directory so stuff isn't cluttered
+    pushd .
+    cd /tmp
 
-  # Install chruby
-  wget -O chruby-0.3.9.tar.gz https://github.com/postmodern/chruby/archive/v0.3.9.tar.gz;
-  tar -xzvf chruby-0.3.9.tar.gz;
-  cd chruby-0.3.9/;
-  sudo make install;
+    # Install chruby
+    wget -O chruby-0.3.9.tar.gz https://github.com/postmodern/chruby/archive/v0.3.9.tar.gz;
+    tar -xzvf chruby-0.3.9.tar.gz;
+    cd chruby-0.3.9/;
+    sudo make install;
 
-  echo "source /usr/local/share/chruby/chruby.sh" >> ~/.bashrc;
-    # add the chruby command to the path
-  echo "source /usr/local/share/chruby/auto.sh" >> ~/.bashrc;
-    # adds auto-switching to chruby
+    echo "source /usr/local/share/chruby/chruby.sh" >> ~/.bashrc;
+      # add the chruby command to the path
+    echo "source /usr/local/share/chruby/auto.sh" >> ~/.bashrc;
+      # adds auto-switching to chruby
+
+    popd
+  fi
 
   source /usr/local/share/chruby/chruby.sh;
   source /usr/local/share/chruby/auto.sh;
-  popd
 fi
 
 chruby ruby-2.3.1
